requirements.txt
diffusers==0.31.0
numpy==2.1.3
Pillow==11.0.0
Requests==2.32.3
torch==2.5.1+cu124
